backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
supabase==2.11.0
python-dotenv==1.0.1
google-auth-oauthlib==1.2.1
google-api-python-client==2.159.0
httpx==0.28.1
python-multipart==0.0.20
yt-dlp>=2024.12.23
